Revert "chore(website): improved api docs (#1421)"
This reverts commit b0f8839de2385f1bc5493509311fb49ded80a889.
