+++ .gitignore
@@ -8,6 +8,3 @@ output
 # Where the runtime testing for Java places node modules
 java_runtime_node
 *.log
-
-# API documentation
-modelina-website/public/docs/api/generated
