+++ modelina-website/src/components/layouts/Footer.tsx
@@ -10,7 +10,7 @@ import Heading from '../typography/Heading';
 
 export default function Footer() {
   return (
-    <footer className="bg-dark margin: 0 auto">
+    <footer className="bg-dark mt-12 margin: 0 auto">
       <div className="max-w-screen-xl mx-auto py-4 sm:py-6 md:py-12 xl:py-16 px-3 overflow-hidden sm:px-6 lg:px-8 divide-y divide-cool-gray">
         <nav className="py-4 sm:py-10 flex flex-wrap justify-between">
           <div className="mr-14 w-full md:w-auto">
