+++ package.json
@@ -46,7 +46,7 @@
     "alterschema": "^1.1.2",
     "change-case": "^4.1.2",
     "openapi-types": "9.3.0",
-    "typescript-json-schema": "^0.57.0"
+    "typescript-json-schema": "^0.53.0"
   },
   "devDependencies": {
     "@asyncapi/parserV1": "npm:@asyncapi/parser@^1.17.2",
@@ -64,12 +64,12 @@
     "eslint-plugin-security": "^1.4.0",
     "eslint-plugin-sonarjs": "^0.10.0",
     "jest": "^27.2.5",
+    "jsdoc-to-markdown": "^7.1.0",
     "markdown-toc": "^1.2.0",
     "puppeteer": "^16.0.0",
     "ts-jest": "^27.0.5",
     "ts-node": "^10.3.0",
-    "typedoc": "^0.24.8",
-    "typescript": "4.9.5"
+    "typescript": "^4.4.4"
   },
   "scripts": {
     "build": "tsc",
@@ -106,11 +106,11 @@
     "generate:runtime:rust": "cross-env CI=true ts-node ./test/runtime/runtime-rust.ts",
     "test:runtime:php": "cross-env CI=true jest ./test/runtime/runtime-php.spec.ts",
     "generate:runtime:php": "cross-env CI=true ts-node ./test/runtime/runtime-php.ts",
-    "test:runtime:typescript": "cd ./test/runtime/runtime-typescript && npm i && npm run test",
+    "test:runtime:typescript": "cross-env CI=true jest ./test/runtime/runtime-typescript",
     "generate:runtime:typescript": "cross-env CI=true ts-node ./test/runtime/runtime-typescript.ts",
     "test:watch": "jest --watch",
-    "docs": "npm run docs:api",
-    "docs:api": "typedoc src/index.ts --out ./modelina-website/public/docs/api/generated --name Modelina",
+    "docs": "npm run docs:markdown",
+    "docs:markdown": "jsdoc2md lib/cjs/index.js -f lib/cjs/**/*.js > API.md",
     "lint": "eslint --max-warnings 0 --config .eslintrc .",
     "lint:fix": "eslint --max-warnings 0 --config .eslintrc . --fix",
     "release": "semantic-release",
