--- test/runtime/runtime-typescript/package.json
@@ -1,12 +1,0 @@
-{
-  "scripts": {
-    "test": "cross-env CI=true jest"
-  },
-  "dependencies": {
-    "jsonbinpack": "1.1.2",
-    "ts-jest": "^27.0.5",
-    "jest": "^27.2.5",
-    "@swc/core": "^1.3.5",
-    "@swc/jest": "^0.2.23"
-  }
-}
